refactor(routines): split prompt composition out of command.rs
command.rs had grown to exactly 500 lines, the line-count gate's ceiling
(linecheck, enforced in .githooks/pre-push and CI), so the next change to
this file — however small — would break CI immediately. Moves prompt
composition and {placeholder} substitution/validation (compose_prompt,
substitute, placeholder_tokens, validate_placeholders,
MAX_INLINE_PROMPT_BYTES, inline_prompt_overflow) into a new
command_prompt.rs, following the same split pattern this file already uses
for command_path_resolution.rs and command_system_prompt.rs. Re-exported via
`pub(crate) use command_prompt::*;` so every existing import path is
unchanged. command.rs drops to 332 lines.

No behavior change.
